Midsem Mini challenge (Threaded Statistics)
This is multithreading of statistic generation.

This use case is a bit contrived though, as in my opinion multithreading is an unnecessary overhead in this case. Almost all the statistics could have been computed with a single pass through the list of numbers. Sorting is rather the bottleneck here 

To run this code, run the compiled code from the terminal and followed by all the numbers needed to compute the statistics separated by spaces as command line arguments